(PC-11746) refactor(QuitIdentityCheckModal): move modal to components
