+++ src/features/identityCheck/pages/Stepper.tsx
@@ -4,8 +4,8 @@ import React, { useEffect } from 'react'
 import styled, { useTheme } from 'styled-components/native'
 
 import { StepButton } from 'features/identityCheck/atoms/StepButton'
+import { QuitIdentityCheckModal } from 'features/identityCheck/components/QuitIdentityCheckModal'
 import { useIdentityCheckContext } from 'features/identityCheck/context/IdentityCheckContextProvider'
-import { QuitIdentityCheckModal } from 'features/identityCheck/pages/QuitIdentityCheckModal'
 import { useIdentityCheckSteps } from 'features/identityCheck/useIdentityCheckSteps'
 import { useGetStepState } from 'features/identityCheck/utils/useGetStepState'
 import { UseNavigationType } from 'features/navigation/RootNavigator'
